Revert "Canonicalize source and target dir"
This reverts commit 13e06459b8aefc1646c93db9dc50a7535a5585ff, which
attempted to ensure prefix stripping does not panic in 'read_source_dir'
when the source is a relative path by canonicalizing it in 'new'. This
commit introduced a new bug by also eagerly canonicalized the target,
which panics if the target dir does not exist (e.g. before the first
build).